Changed variable names to our convention, changed "Command Swerve Drivetrain" to "SwerveSubsystem" and moved it to the subsystem tab.
